+++ packages/core/src/hooks/data/useUpdate.ts
@@ -34,7 +34,7 @@ import {
     useRefineContext,
 } from "@hooks";
 import {
-    queryKeys,
+    queryKeysReplacement,
     pickDataProvider,
     pickNotDeprecated,
     useActiveAuthProvider,
@@ -44,6 +44,7 @@ import {
     UseLoadingOvertimeOptionsProps,
     UseLoadingOvertimeReturnType,
 } from "../useLoadingOvertime";
+import { useKeys } from "@hooks/useKeys";
 
 export type UpdateParams<TData, TError, TVariables> = {
     /**
@@ -168,6 +169,7 @@ export const useUpdate = <
     const {
         options: { textTransformers },
     } = useRefineContext();
+    const { keys, preferLegacyKeys } = useKeys();
 
     const mutation = useMutation<
         UpdateResponse<TData>,
@@ -268,20 +270,32 @@ export const useUpdate = <
 
                 const preferredMeta = pickNotDeprecated(meta, metaData);
 
-                const queryKey = queryKeys(
+                const queryKey = queryKeysReplacement(preferLegacyKeys)(
                     identifier,
                     pickDataProvider(identifier, dataProviderName, resources),
                     preferredMeta,
                 );
 
+                const resourceKeys = keys()
+                    .data(
+                        pickDataProvider(
+                            identifier,
+                            dataProviderName,
+                            resources,
+                        ),
+                    )
+                    .resource(identifier);
+
                 const previousQueries: PreviousQuery<TData>[] =
-                    queryClient.getQueriesData(queryKey.resourceAll);
+                    queryClient.getQueriesData(
+                        resourceKeys.get(preferLegacyKeys),
+                    );
 
                 const mutationModePropOrContext =
                     mutationMode ?? mutationModeContext;
 
                 await queryClient.cancelQueries(
-                    queryKey.resourceAll,
+                    resourceKeys.get(preferLegacyKeys),
                     undefined,
                     {
                         silent: true,
@@ -291,7 +305,10 @@ export const useUpdate = <
                 if (mutationModePropOrContext !== "pessimistic") {
                     // Set the previous queries to the new ones:
                     queryClient.setQueriesData(
-                        queryKey.list(),
+                        resourceKeys
+                            .action("list")
+                            .params(preferredMeta ?? {})
+                            .get(preferLegacyKeys),
                         (previous?: GetListResponse<TData> | null) => {
                             if (!previous) {
                                 return null;
@@ -315,7 +332,7 @@ export const useUpdate = <
                     );
 
                     queryClient.setQueriesData(
-                        queryKey.many(),
+                        resourceKeys.action("many").get(preferLegacyKeys),
                         (previous?: GetListResponse<TData> | null) => {
                             if (!previous) {
                                 return null;
@@ -339,7 +356,11 @@ export const useUpdate = <
                     );
 
                     queryClient.setQueriesData(
-                        queryKey.detail(id),
+                        resourceKeys
+                            .action("one")
+                            .id(id)
+                            .params(preferredMeta ?? {})
+                            .get(preferLegacyKeys),
                         (previous?: GetListResponse<TData> | null) => {
                             if (!previous) {
                                 return null;
@@ -516,6 +537,7 @@ export const useUpdate = <
                     });
                 }
             },
+            mutationKey: keys().data().mutation("update").get(preferLegacyKeys),
             ...mutationOptions,
         },
     );