+++ packages/core/src/hooks/invalidate/index.tsx
@@ -2,8 +2,9 @@ import { useCallback } from "react";
 import { useQueryClient } from "@tanstack/react-query";
 
 import { useResource } from "@hooks/resource";
-import { queryKeys, pickDataProvider } from "@definitions";
+import { pickDataProvider } from "@definitions";
 import { BaseKey, IQueryKeys } from "../../interfaces";
+import { useKeys } from "@hooks/useKeys";
 
 export type UseInvalidateProp = {
     resource?: string;
@@ -15,6 +16,7 @@ export type UseInvalidateProp = {
 export const useInvalidate = (): ((props: UseInvalidateProp) => void) => {
     const { resources } = useResource();
     const queryClient = useQueryClient();
+    const { keys, preferLegacyKeys } = useKeys();
 
     const invalidate = useCallback(
         ({
@@ -26,28 +28,40 @@ export const useInvalidate = (): ((props: UseInvalidateProp) => void) => {
             if (invalidates === false) {
                 return;
             }
-            const queryKey = queryKeys(
-                resource,
-                pickDataProvider(resource, dataProviderName, resources),
-            );
+            const dp = pickDataProvider(resource, dataProviderName, resources);
+
+            const queryKey = keys()
+                .data(dp)
+                .resource(resource ?? "");
 
             invalidates.forEach((key) => {
                 switch (key) {
                     case "all":
-                        queryClient.invalidateQueries(queryKey.all);
+                        queryClient.invalidateQueries(
+                            keys().data(dp).get(preferLegacyKeys),
+                        );
                         break;
                     case "list":
-                        queryClient.invalidateQueries(queryKey.list());
+                        queryClient.invalidateQueries(
+                            queryKey.action("list").get(preferLegacyKeys),
+                        );
                         break;
                     case "many":
-                        queryClient.invalidateQueries(queryKey.many());
+                        queryClient.invalidateQueries(
+                            queryKey.action("many").get(preferLegacyKeys),
+                        );
                         break;
                     case "resourceAll":
-                        queryClient.invalidateQueries(queryKey.resourceAll);
+                        queryClient.invalidateQueries(
+                            queryKey.get(preferLegacyKeys),
+                        );
                         break;
                     case "detail":
                         queryClient.invalidateQueries(
-                            queryKey.detail(id || ""),
+                            queryKey
+                                .action("one")
+                                .id(id || "")
+                                .get(preferLegacyKeys),
                         );
                         break;
                     default: